+++ src/fields/AddressField.php
@@ -20,12 +20,14 @@ use craft\elements\Entry;
 use craft\helpers\Json;
 use doublesecretagency\googlemaps\enums\Defaults;
 use doublesecretagency\googlemaps\GoogleMapsPlugin;
+use doublesecretagency\googlemaps\gql\types\Address as AddressType;
 use doublesecretagency\googlemaps\helpers\ProximitySearchHelper;
 use doublesecretagency\googlemaps\models\Address as AddressModel;
 use doublesecretagency\googlemaps\records\Address as AddressRecord;
 use doublesecretagency\googlemaps\validators\AddressValidator;
 use doublesecretagency\googlemaps\web\assets\AddressFieldAsset;
 use doublesecretagency\googlemaps\web\assets\AddressFieldSettingsAsset;
+use GraphQL\Type\Definition\Type;
 
 /**
  * Class AddressField
@@ -529,6 +531,14 @@ class AddressField extends Field implements PreviewableFieldInterface
 
         // Modify the element query to perform a proximity search
         ProximitySearchHelper::modifyElementsQuery($query, $value, $this);
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getContentGqlType(): Type|array
+    {
+        return AddressType::getType();
     }
 
 }